+++ src/main/java/urecagroup1backend/meeting_room/dto/ReservationRequest.java
@@ -1,4 +1,4 @@
-package urecagroup1backend.meeting_room.api.dto;
+package urecagroup1backend.meeting_room.dto;
 
 public record ReservationRequest(
         String phoneNumber
